maj modify + delete and like/dislike

+++ controllers/sauce.js
@@ -47,8 +47,10 @@ exports.modifySauce = (req, res, next) => {
             ...JSON.parse(req.body.sauce),
             imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         } : { ...req.body }
+    delete sauceObject.userId
     Sauce.findOne({ _id: req.params.id, userId: req.auth.userId })
         .then(sauce => {
+            if (!sauce) return res.status(403).json({ message: 'Vous n\'avez pas l\'autorisation' })
             if (req.file) {
                 const filename = sauce.imageUrl.replace(`${req.protocol}://${req.get('host')}`, "")
                 fs.unlinkSync(__dirname + '/..' + filename)
@@ -57,12 +59,12 @@ exports.modifySauce = (req, res, next) => {
                 .then(() => res.status(201).json({ message: 'Objet modifié !' }))
                 .catch(error => {
                     if (req.file) fs.unlinkSync(__dirname + '/../images/' + req.file.filename)
-                    res.status(403).json({ error })
+                    return res.status(403).json({ error })
                 });
         })
         .catch(error => {
             if (req.file) fs.unlinkSync(__dirname + '/../images/' + req.file.filename)
-            res.status(403).json({ error })
+            return res.status(403).json({ error })
         });
 }
 
@@ -77,31 +79,38 @@ exports.deleteSauce = (req, res, next) => {
                     .catch(error => res.status(400).json({ error }));
             })
         })
-        .catch(error => res.status(500).json({ error }))
+        .catch(error => res.status(403).json({ error }))
 }
 
 // Liker sauce ou Dislike
 exports.likeSauce = (req, res, next) => {
     if (req.body.like === 1) {
-        Sauce.updateOne({ _id: req.params.id, usersLiked: { $ne: req.auth.userId } }, { $inc: { likes: +1 }, $push: { usersLiked: req.auth.userId } })
-            .then(() => res.status(200).json({ message: 'Like ajouté !' }))
+        Sauce.updateOne({ _id: req.params.id, usersLiked: { $ne: req.auth.userId }, usersDisliked: { $ne: req.auth.userId } }, { $inc: { likes: +1 }, $push: { usersLiked: req.auth.userId } })
+            .then((result) => {
+                if (result.modifiedCount === 0) return res.status(403).json({ message: 'Vous avez déjà like ou dislike la sauce' })
+                return res.status(200).json({ message: 'Like ajouté !' })
+            })
             .catch(error => res.status(400).json({ error }))
     } else if (req.body.like === -1) {
-        Sauce.updateOne({ _id: req.params.id, usersDisliked: { $ne: req.auth.userId } }, { $inc: { dislikes: +1 }, $push: { usersDisliked: req.auth.userId } })
-            .then(() => res.status(200).json({ message: 'Dislike ajouté !' }))
+        Sauce.updateOne({ _id: req.params.id, usersLiked: { $ne: req.auth.userId }, usersDisliked: { $ne: req.auth.userId } }, { $inc: { dislikes: +1 }, $push: { usersDisliked: req.auth.userId } })
+            .then((result) => {
+                if (result.modifiedCount === 0) return res.status(403).json({ message: 'Vous avez déjà like ou dislike la sauce' })
+                return res.status(200).json({ message: 'Dislike ajouté !' })
+            })
             .catch(error => res.status(400).json({ error }))
     } else {
         Sauce.findOne({ _id: req.params.id })
             .then(sauce => {
                 if (sauce.usersLiked.includes(req.auth.userId)) {
                     Sauce.updateOne({ _id: req.params.id }, { $pull: { usersLiked: req.auth.userId }, $inc: { likes: -1 } })
-                        .then(() => { res.status(200).json({ message: 'Like supprimé !' }) })
+                        .then(() => res.status(200).json({ message: 'Like supprimé !' }))
                         .catch(error => res.status(400).json({ error }))
                 } else if (sauce.usersDisliked.includes(req.auth.userId)) {
                     Sauce.updateOne({ _id: req.params.id }, { $pull: { usersDisliked: req.auth.userId }, $inc: { dislikes: -1 } })
-                        .then(() => { res.status(200).json({ message: 'Dislike supprimé !' }) })
+                        .then(() => res.status(200).json({ message: 'Dislike supprimé !' }))
                         .catch(error => res.status(400).json({ error }))
                 }
+                else return res.status(403).json({ message: 'Vous n\'avez pas like ou dislike la sauce' })
             })
             .catch(error => res.status(400).json({ error }))
     }
